Moved `phasor_filter` to `phasor` module. Updated typing and documentation for `phasor_filter`.
